+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Automatically simplfies argv list for Hydra experiments when same parameter is r
 Better error message when making a typo in the grid name. Always show the traceback when getting an
 import error.
 
+Added import/export command to easily share XP hyper-params in text form.
+
 ## [0.1.7] - 2021-11-08
 
 Adding support for type arrays.