+++ CHANGELOG.md
@@ -22,6 +22,11 @@ import error.
 
 Added import/export command to easily share XP hyper-params in text form.
 
+Added shared repository option (`shared` option in Dora config). No metrics or
+checkpoints can be shared, this is still a bit dangerous, but this will act as a shared
+database for mappings from SIG -> hyper params, so that you can just pass a SIG
+to your teammate and launch the same XP.
+
 ## [0.1.7] - 2021-11-08
 
 Adding support for type arrays.