+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.9a]
+
+Reliable rmtree used to avoid `--clear` being blocked by some locking issues on NFS.
+
 ## [0.1.8] - 2021-12-30
 
 Always export RANK and WORLD_SIZE as env variable, so that they can be consumed by Hydra config